Extract room_member_count_is.rs
